requirements.txt
pillow==10.4.0
pillow-heif==1.1.1
gradio==3.41.2
tqdm
colorama

